finish refactoring base settings page
