+++ src/components/PageHeader/types.ts
@@ -1,0 +1,6 @@
+import { ReactNode } from 'react';
+
+export interface IPageHeaderProps {
+  title: ReactNode;
+  children?: ReactNode;
+}
